Rename droughtRegions as droughtRisks
