+++ frontend/src/config/droughtRisks/data-formats.ts
