+++ frontend/src/config/droughtRisks/metadata.ts
