+++ frontend/src/config/droughtRisks/styles.ts
